Clean up and remove unused JS
