--- cfgov/unprocessed/apps/find-a-housing-counselor/js/main.js
@@ -1,2 +1,0 @@
-const hud = require( './hud' );
-hud.initializeMap();
